feat: refactored match to respect clean arch
