+++ src/main/java/com/luizmatias/findadev/api/dtos/MatchDTO.java
@@ -1,0 +1,12 @@
+package com.luizmatias.findadev.api.dtos;
+
+import java.time.LocalDateTime;
+
+public record MatchDTO(
+        Long id,
+        UserDTO clientUser,
+        UserDTO developerUser,
+        LocalDateTime matchedAt,
+        LocalDateTime unmatchedAt
+) {
+}
